Refactor parser class under chain of responsibility design pattern Edit tests to fit new parser definition Comment out ParserTest as it is no longer relevant
